+++ lib/world/blocks.ts
@@ -29,7 +29,8 @@ export const enum BlockId {
   WheatStage0 = 22,
   WheatStage1 = 23,
   WheatStage2 = 24,
-  WheatStage3 = 25
+  WheatStage3 = 25,
+  Furnace = 26
 }
 
 export enum BiomeId {
@@ -59,7 +60,8 @@ export const HELD_BLOCK_COLORS: Partial<Record<BlockId, number>> = {
   [BlockId.DiamondOre]: 0x85e9f4,
   [BlockId.Snow]: 0xf2f5fa,
   [BlockId.Cactus]: 0x6aa850,
-  [BlockId.Bed]: 0xc0392b
+  [BlockId.Bed]: 0xc0392b,
+  [BlockId.Furnace]: 0x63666a
 };
 
 export const HELD_BLOCK_FALLBACK_COLOR = 0xbababa;
@@ -92,5 +94,6 @@ export const BLOCK_COLORS: Record<number, [number, number, number]> = {
   [BlockId.WheatStage0]: [0.4, 0.62, 0.25],
   [BlockId.WheatStage1]: [0.55, 0.66, 0.27],
   [BlockId.WheatStage2]: [0.7, 0.69, 0.29],
-  [BlockId.WheatStage3]: [0.82, 0.72, 0.3]
+  [BlockId.WheatStage3]: [0.82, 0.72, 0.3],
+  [BlockId.Furnace]: [0.38, 0.39, 0.41]
 };